Firmware: added smooth slowdown and brightness support;

+++ Firmware/Lightpack.h
@@ -55,7 +55,7 @@
 // Global variables
 extern Settings_t g_Settings;
 extern Images_t g_Images;
-extern volatile uint8_t g_smoothIndex;
+extern volatile uint16_t g_smoothIndex;
 
 #endif /* LIGHTPACK_H_INCLUDED */
